% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{process_fun2}
\alias{process_fun2}
\title{QoI wrapper for the process model}
\usage{
process_fun2(X, M = 50, nrep = 10)
}
\arguments{
\item{X}{Matrix or data.frame of parameters.}

\item{M}{Target number of successes.}

\item{nrep}{Number of repetitions for the QoI.}
}
\value{
Numeric vector of QoI values.
}
\description{
For each row of \code{X}, evaluates \code{process_fun1} several times
and returns the mean time to M successes.
}
