% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{sobol4r_g_function}
\alias{sobol4r_g_function}
\title{Sobol g function (Saltelli) for arbitrary dimension}
\usage{
sobol4r_g_function(
  X,
  a = c(0, 1, 4.5, 9, 99, 99, 99, 99),
  use_cols = seq_len(ncol(as.data.frame(X)))
)
}
\arguments{
\item{X}{Numeric matrix or data.frame of inputs in \link{0,1}.
Rows are observations, columns are factors.}

\item{a}{Numeric vector of non negative parameters (importance factors).
Its length must be at least the number of columns used.}

\item{use_cols}{Integer vector giving which columns of X to use.
Defaults to all columns.}
}
\value{
Numeric vector of length nrow(X) with the model output.
}
\description{
Generic implementation of the non monotonic Sobol g function.
}
