% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{sobol4r_one_unit}
\alias{sobol4r_one_unit}
\title{One unit of the simple stochastic process}
\usage{
sobol4r_one_unit(lambda1, lambda2, lambda3, p1, p2)
}
\arguments{
\item{lambda1}{Rate of the first exponential time.}

\item{lambda2}{Rate of the second exponential time.}

\item{lambda3}{Rate of the third exponential time.}

\item{p1}{First Bernoulli success probability.}

\item{p2}{Second Bernoulli success probability.}
}
\value{
Numeric vector of length 3:
\itemize{
\item indicator of success (0 or 1)
\item time to first event
\item additional time if success (0 otherwise)
}
}
\description{
This is the \code{one_unit} function from the script.
}
