% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{sobol4r_process_indiv}
\alias{sobol4r_process_indiv}
\title{Time to M successes for one individual}
\usage{
sobol4r_process_indiv(X_indiv, M = 50L)
}
\arguments{
\item{X_indiv}{Numeric vector of length 5:
(lambda1, lambda2, lambda3, p1, p2).}

\item{M}{Target number of successes.}
}
\value{
Time to reach M successes for that individual.
The value has an attribute "success" with the final number
of successes (should be >= M).
}
\description{
This is the \code{process.fun_indiv} function from the script.
}
