% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_g_wrappers.R
\name{sobol_g2_additive_noise}
\alias{sobol_g2_additive_noise}
\title{Additive Gaussian noise on the Sobol G-function (k = 2) - C++ backend}
\usage{
sobol_g2_additive_noise(X, sd = 1, a = c(0, 1, 4.5, 9, 99, 99, 99, 99))
}
\arguments{
\item{X}{Numeric matrix or data.frame with at least two columns.}

\item{sd}{Standard deviation of the Gaussian noise.}

\item{a}{Numeric vector of parameters (at least length 2).}
}
\value{
Numeric vector of model outputs with noise.
}
\description{
Additive Gaussian noise on the Sobol G-function (k = 2) - C++ backend
}
