% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol4r_g_functions.R
\name{sobol_g2_with_covariate_noise_R}
\alias{sobol_g2_with_covariate_noise_R}
\title{Additive Gaussian noise on the Sobol G-function (k = 2)}
\usage{
sobol_g2_with_covariate_noise_R(X)
}
\arguments{
\item{X}{Numeric matrix or data.frame with at least two columns.}
}
\value{
Numeric vector of model outputs with noise.
}
\description{
Additive Gaussian noise on the Sobol G-function (k = 2)
}
