% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.sobol_result}
\alias{autoplot.sobol_result}
\alias{autoplot.sobol}
\alias{autoplot.sobol2007}
\alias{autoplot.sobol_summary}
\title{#' @export
autoplot <- function(object, ...) {
UseMethod("autoplot")
}
Autoplot for \code{sobol_result} class}
\usage{
autoplot.sobol_result(
  object,
  show_uncertainty = FALSE,
  probs = c(0.1, 0.9),
  bootstrap = 200L,
  ...
)

autoplot.sobol(object, separate_panels = TRUE, ncol = 2, ...)

autoplot.sobol2007(object, ...)

autoplot.sobol_summary(object, ...)
}
\description{
#' @export
autoplot <- function(object, ...) {
UseMethod("autoplot")
}
Autoplot for \code{sobol_result} class

Autoplot for \code{sobol} class

Autoplot for \code{sobol2007} class

Autoplot for \code{sobol_summary} class
}
