% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{sobol_example_covariate_large}
\alias{sobol_example_covariate_large}
\title{Example 3: Large covariate dependent random effect}
\usage{
sobol_example_covariate_large(
  n = 50000,
  nrep_qoi = 1000,
  order = 2,
  nboot = 100
)
}
\arguments{
\item{n}{Monte Carlo sample size for each base design.}

\item{nrep_qoi}{Number of repetitions for the QoI.}

\item{order}{Maximum interaction order.}

\item{nboot}{Number of bootstrap replicates.}
}
\value{
A list with two \code{"sobol"} objects:
\code{x_single} (single noisy run),
\code{x_qoi} (QoI-based indices).
}
\description{
Third input C3 is uniform on [1, 100], used as the mean of a Gaussian
noise term added to the G-function. Quantity of interest is the mean
of repeated evaluations.
}
