% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobol4R_functions.R
\name{sobol_example_covariate_small}
\alias{sobol_example_covariate_small}
\title{Example 4: Slight covariate dependent random effect}
\usage{
sobol_example_covariate_small(
  n = 50000,
  nrep_qoi = 1000,
  order = 2,
  nboot = 100
)
}
\arguments{
\item{n}{Monte Carlo sample size for each base design.}

\item{nrep_qoi}{Number of repetitions for the QoI.}

\item{order}{Maximum interaction order.}

\item{nboot}{Number of bootstrap replicates.}
}
\value{
A list with two \code{"sobol"} objects:
\code{x_single} (single noisy run),
\code{x_qoi} (QoI-based indices).
}
\description{
Same as \code{sobol_example_covariate_large} but with C3 uniform
on [1, 1.5], that is with a much smaller range for the mean of the
Gaussian noise.
}
