% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_g_wrappers.R
\name{sobol_g2_with_covariate_noise}
\alias{sobol_g2_with_covariate_noise}
\title{Covariate dependent Gaussian noise on the Sobol G-function (k = 2) - C++ backend}
\usage{
sobol_g2_with_covariate_noise(X, a = c(0, 1, 4.5, 9, 99, 99, 99, 99))
}
\arguments{
\item{X}{Numeric matrix or data.frame with at least two columns.}

\item{a}{Numeric vector of parameters (at least length 2).}
}
\value{
Numeric vector of model outputs with noise.
}
\description{
Covariate dependent Gaussian noise on the Sobol G-function (k = 2) - C++ backend
}
