% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_g_functions.R
\name{sobol_g_function_R}
\alias{sobol_g_function_R}
\title{Sobol G-function (Saltelli reference function)}
\usage{
sobol_g_function_R(X, a = c(0, 1, 4.5, 9, 99, 99, 99, 99))
}
\arguments{
\item{X}{Numeric matrix or data.frame of inputs in [0, 1].}

\item{a}{Numeric vector of parameters a_j controlling importance.
Its length must be at least \code{ncol(X)}.}
}
\value{
Numeric vector of length \code{nrow(X)} with model outputs.
}
\description{
Generic implementation of the Sobol G-function for k inputs.
Columns of \code{X} are interpreted as inputs X1, X2, ..., Xk.
}
